DotRocket: return double instead of float

+++ contrib/DotRocket/DotRocket/DotRocket.cpp
@@ -15,7 +15,7 @@ private ref class PlayerTrack: public Track {
 	const sync_track *track;
 public:
 	PlayerTrack(const sync_track *track): track(track) {}
-	virtual float GetValue(double time) override
+	virtual double GetValue(double time) override
 	{
 		return sync_get_val(track, time);
 	};
